refactor(krites/parse): full structural overhaul — snafu errors, quality (#3218)
Comprehensive overhaul of the 4.5K LOC Datalog parser module:

- Split monoliths: expr.rs (623L) → expr/{mod,bytecode,strings}.rs;
  sys/parse.rs (1001L) → sys/{parse,index}.rs;
  query/program.rs (681L) → query/{program,options}.rs
- Enriched ParseError with UnexpectedRule, MissingElement, InvalidQueryAt,
  InvalidInteger variants carrying SourceSpan for diagnostics
- DRY: unified string escape handling (was duplicated for quoted/single-quoted),
  extracted tokenizer/filter parsing shared by FTS/LSH/HNSW indexes,
  consolidated 4x duplicated index_drop parsing into single helper,
  factored sysop write-lock collection with insert_index_lock helper
- Fixed parse_int to return Result instead of panicking via unreachable!(),
  replaced unreachable!() in collect_sort_option with debug_assert
- Added input_location_to_span and unexpected_rule_error helpers to mod.rs
- Every function now has accurate documentation comments
- Added //! module docs to all 16 files
- Added #[must_use] on SourceSpan::merge and input_location_to_span
- Added doc comments on all public enum variants (DatalogScript, ImperativeStmt,
  SysOp, HnswDistance) and struct fields
- Removed stale "Datalog query parsing." duplicate module docs
- No behavioral changes: all 328 tests pass, zero clippy warnings in parse/
