+++ crates/krites/src/parse/error.rs
@@ -7,6 +7,8 @@
 //! allowing `?`-based propagation into engine-internal `InternalResult<T>` contexts.
 use snafu::Snafu;
 
+use super::SourceSpan;
+
 /// Structured error for all Datalog parse failures.
 #[derive(Debug, Snafu)]
 #[snafu(visibility(pub(crate)))]
@@ -29,10 +31,50 @@ pub enum ParseError {
         location: snafu::Location,
     },
 
+    /// Semantically invalid query structure with source location.
+    #[snafu(display("invalid query at {span}: {message}"))]
+    InvalidQueryAt {
+        span: SourceSpan,
+        message: String,
+        #[snafu(implicit)]
+        location: snafu::Location,
+    },
+
     /// Reference to a variable that is not bound in any rule head.
     #[snafu(display("unbound variable '{name}'"))]
     UnboundVariable {
         name: String,
+        #[snafu(implicit)]
+        location: snafu::Location,
+    },
+
+    /// A grammar rule appeared where it was not expected.
+    ///
+    /// This indicates a mismatch between the pest grammar and the parser
+    /// implementation. If encountered, file a bug report.
+    #[snafu(display("unexpected grammar rule {rule:?} at {span} in {context}"))]
+    UnexpectedRule {
+        rule: String,
+        span: SourceSpan,
+        context: &'static str,
+        #[snafu(implicit)]
+        location: snafu::Location,
+    },
+
+    /// A required child element was missing from a pest parse pair.
+    #[snafu(display("missing {element} at {span}"))]
+    MissingElement {
+        element: &'static str,
+        span: SourceSpan,
+        #[snafu(implicit)]
+        location: snafu::Location,
+    },
+
+    /// Integer literal could not be parsed.
+    #[snafu(display("invalid integer literal at {span}: {message}"))]
+    InvalidInteger {
+        span: SourceSpan,
+        message: String,
         #[snafu(implicit)]
         location: snafu::Location,
     },
