--- crates/krites/src/parse/expr.rs
@@ -1,623 +1,0 @@
-//! Expression parsing from Datalog source.
-#![expect(
-    clippy::as_conversions,
-    clippy::indexing_slicing,
-    clippy::needless_return,
-    clippy::pedantic,
-    clippy::result_large_err,
-    reason = "Datalog expression parser — numeric casts for Unicode escapes, indexing into pest pairs by structural position, InternalError is the crate-wide Result type"
-)]
-
-use std::collections::BTreeMap;
-use std::sync::LazyLock;
-
-use compact_str::CompactString;
-use itertools::Itertools;
-use pest::pratt_parser::{Op, PrattParser};
-
-use crate::data::expr::{Bytecode, Expr, get_op, no_implementation_err};
-use crate::data::functions::{
-    OP_ADD, OP_AND, OP_COALESCE, OP_CONCAT, OP_DIV, OP_EQ, OP_GE, OP_GT, OP_JSON_OBJECT, OP_LE,
-    OP_LIST, OP_LT, OP_MAYBE_GET, OP_MINUS, OP_MOD, OP_MUL, OP_NEGATE, OP_NEQ, OP_OR, OP_POW,
-    OP_SUB,
-};
-use crate::data::symb::Symbol;
-use crate::data::value::DataValue;
-use crate::error::InternalResult as Result;
-use crate::parse::error::InvalidQuerySnafu;
-use crate::parse::{ExtractSpan, Pair, Rule, SourceSpan};
-
-static PRATT_PARSER: LazyLock<PrattParser<Rule>> = LazyLock::new(|| {
-    use pest::pratt_parser::Assoc::*;
-
-    PrattParser::new()
-        .op(Op::infix(Rule::op_or, Left))
-        .op(Op::infix(Rule::op_and, Left))
-        .op(Op::infix(Rule::op_gt, Left)
-            | Op::infix(Rule::op_lt, Left)
-            | Op::infix(Rule::op_ge, Left)
-            | Op::infix(Rule::op_le, Left))
-        .op(Op::infix(Rule::op_eq, Left) | Op::infix(Rule::op_ne, Left))
-        .op(Op::infix(Rule::op_mod, Left))
-        .op(Op::infix(Rule::op_add, Left)
-            | Op::infix(Rule::op_sub, Left)
-            | Op::infix(Rule::op_concat, Left))
-        .op(Op::infix(Rule::op_mul, Left) | Op::infix(Rule::op_div, Left))
-        .op(Op::infix(Rule::op_pow, Right))
-        .op(Op::infix(Rule::op_coalesce, Left))
-        .op(Op::prefix(Rule::minus))
-        .op(Op::prefix(Rule::negate))
-        .op(Op::infix(Rule::op_field_access, Left))
-});
-
-pub(crate) fn expr2bytecode(expr: &Expr, collector: &mut Vec<Bytecode>) -> Result<()> {
-    match expr {
-        Expr::Binding { var, tuple_pos } => collector.push(Bytecode::Binding {
-            var: var.clone(),
-            tuple_pos: *tuple_pos,
-        }),
-        Expr::Const { val, span } => collector.push(Bytecode::Const {
-            val: val.clone(),
-            span: *span,
-        }),
-        Expr::Apply { op, args, span } => {
-            let arity = args.len();
-            for arg in args.iter() {
-                expr2bytecode(arg, collector)?;
-            }
-            collector.push(Bytecode::Apply {
-                op,
-                arity,
-                span: *span,
-            })
-        }
-        Expr::Cond { clauses, span } => {
-            let mut return_jump_pos = vec![];
-            for (cond, val) in clauses {
-                expr2bytecode(cond, collector)?;
-                collector.push(Bytecode::JumpIfFalse {
-                    jump_to: 0,
-                    span: *span,
-                });
-                let false_jump_amend_pos = collector.len() - 1;
-                expr2bytecode(val, collector)?;
-                collector.push(Bytecode::Goto {
-                    jump_to: 0,
-                    span: *span,
-                });
-                return_jump_pos.push(collector.len() - 1);
-                // SAFETY: `false_jump_amend_pos` is `collector.len() - 1` from above,
-                // so it is always a valid index.
-                collector[false_jump_amend_pos] = Bytecode::JumpIfFalse {
-                    jump_to: collector.len(),
-                    span: *span,
-                };
-            }
-            let total_len = collector.len();
-            for pos in return_jump_pos {
-                // SAFETY: `pos` values are `collector.len() - 1` values pushed above,
-                // so they are always valid indices.
-                collector[pos] = Bytecode::Goto {
-                    jump_to: total_len,
-                    span: *span,
-                }
-            }
-        }
-        Expr::UnboundApply { op, .. } => {
-            return Err(no_implementation_err(op).into());
-        }
-    }
-    Ok(())
-}
-
-pub(crate) fn build_expr(pair: Pair<'_>, param_pool: &BTreeMap<String, DataValue>) -> Result<Expr> {
-    if pair.as_rule() != Rule::expr {
-        return Err(InvalidQuerySnafu {
-            message: "Invalid expression encountered".to_string(),
-        }
-        .build()
-        .into());
-    }
-
-    PRATT_PARSER
-        .map_primary(|v| build_term(v, param_pool))
-        .map_infix(build_expr_infix)
-        .map_prefix(|op, rhs| {
-            let rhs = rhs?;
-            let rhs_span = rhs.span();
-            Ok(match op.as_rule() {
-                Rule::minus => Expr::Apply {
-                    op: &OP_MINUS,
-                    args: [rhs].into(),
-                    span: op.extract_span().merge(rhs_span),
-                },
-                Rule::negate => Expr::Apply {
-                    op: &OP_NEGATE,
-                    args: [rhs].into(),
-                    span: op.extract_span().merge(rhs_span),
-                },
-                r => {
-                    return Err(InvalidQuerySnafu {
-                        message: format!(
-                            "unexpected rule {:?} in parser - grammar mismatch, please file a bug",
-                            r
-                        ),
-                    }
-                    .build()
-                    .into());
-                }
-            })
-        })
-        .parse(pair.into_inner())
-}
-
-fn build_expr_infix(lhs: Result<Expr>, op: Pair<'_>, rhs: Result<Expr>) -> Result<Expr> {
-    let args = vec![lhs?, rhs?];
-    let op = match op.as_rule() {
-        Rule::op_add => &OP_ADD,
-        Rule::op_sub => &OP_SUB,
-        Rule::op_mul => &OP_MUL,
-        Rule::op_div => &OP_DIV,
-        Rule::op_mod => &OP_MOD,
-        Rule::op_pow => &OP_POW,
-        Rule::op_eq => &OP_EQ,
-        Rule::op_ne => &OP_NEQ,
-        Rule::op_gt => &OP_GT,
-        Rule::op_ge => &OP_GE,
-        Rule::op_lt => &OP_LT,
-        Rule::op_le => &OP_LE,
-        Rule::op_concat => &OP_CONCAT,
-        Rule::op_or => &OP_OR,
-        Rule::op_and => &OP_AND,
-        Rule::op_coalesce => &OP_COALESCE,
-        Rule::op_field_access => &OP_MAYBE_GET,
-        r => {
-            return Err(InvalidQuerySnafu {
-                message: format!(
-                    "unexpected rule {:?} in parser - grammar mismatch, please file a bug",
-                    r
-                ),
-            }
-            .build()
-            .into());
-        }
-    };
-    #[expect(clippy::indexing_slicing, reason = "index bounds validated")]
-    let start = args[0].span().0;
-    #[expect(clippy::indexing_slicing, reason = "index bounds validated")]
-    let end = args[1].span().0 + args[1].span().1;
-    let length = end - start;
-    Ok(Expr::Apply {
-        op,
-        args: args.into(),
-        span: SourceSpan(start, length),
-    })
-}
-
-fn build_term(pair: Pair<'_>, param_pool: &BTreeMap<String, DataValue>) -> Result<Expr> {
-    let span = pair.extract_span();
-    let op = pair.as_rule();
-    Ok(match op {
-        Rule::var => Expr::Binding {
-            var: Symbol::new(pair.as_str(), pair.extract_span()),
-            tuple_pos: None,
-        },
-        Rule::param => {
-            let param_str = pair.as_str().strip_prefix('$').ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "parameter missing '$' prefix".to_string(),
-                }
-                .build()
-            })?;
-            Expr::Const {
-                val: param_pool
-                    .get(param_str)
-                    .ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: format!("Required parameter {param_str} not found"),
-                        }
-                        .build()
-                    })?
-                    .clone(),
-                span,
-            }
-        }
-        Rule::pos_int => {
-            let i = pair.as_str().replace('_', "").parse::<i64>().map_err(|e| {
-                InvalidQuerySnafu {
-                    message: format!("Cannot parse integer: {e}"),
-                }
-                .build()
-            })?;
-            Expr::Const {
-                val: DataValue::from(i),
-                span,
-            }
-        }
-        Rule::hex_pos_int => {
-            let i = parse_int(pair.as_str(), 16);
-            Expr::Const {
-                val: DataValue::from(i),
-                span,
-            }
-        }
-        Rule::octo_pos_int => {
-            let i = parse_int(pair.as_str(), 8);
-            Expr::Const {
-                val: DataValue::from(i),
-                span,
-            }
-        }
-        Rule::bin_pos_int => {
-            let i = parse_int(pair.as_str(), 2);
-            Expr::Const {
-                val: DataValue::from(i),
-                span,
-            }
-        }
-        Rule::dot_float | Rule::sci_float => {
-            let f = pair.as_str().replace('_', "").parse::<f64>().map_err(|e| {
-                InvalidQuerySnafu {
-                    message: format!("Cannot parse float: {e}"),
-                }
-                .build()
-            })?;
-            Expr::Const {
-                val: DataValue::from(f),
-                span,
-            }
-        }
-        Rule::null => Expr::Const {
-            val: DataValue::Null,
-            span,
-        },
-        Rule::boolean => Expr::Const {
-            val: DataValue::from(pair.as_str() == "true"),
-            span,
-        },
-        Rule::quoted_string | Rule::s_quoted_string | Rule::raw_string => {
-            let s = parse_string(pair)?;
-            Expr::Const {
-                val: DataValue::Str(s),
-                span,
-            }
-        }
-        Rule::list => {
-            let mut collected = vec![];
-            for p in pair.into_inner() {
-                collected.push(build_expr(p, param_pool)?)
-            }
-            Expr::Apply {
-                op: &OP_LIST,
-                args: collected.into(),
-                span,
-            }
-        }
-        Rule::object => {
-            let mut args = vec![];
-            for p in pair.into_inner() {
-                let mut p = p.into_inner();
-                let k = p.next().ok_or_else(|| {
-                    InvalidQuerySnafu {
-                        message: "missing key in object pair".to_string(),
-                    }
-                    .build()
-                })?;
-                let v = p.next().ok_or_else(|| {
-                    InvalidQuerySnafu {
-                        message: "missing value in object pair".to_string(),
-                    }
-                    .build()
-                })?;
-                let k = build_expr(k, param_pool)?;
-                let v = build_expr(v, param_pool)?;
-                args.push(k);
-                args.push(v);
-            }
-            Expr::Apply {
-                op: &OP_JSON_OBJECT,
-                args: args.into(),
-                span,
-            }
-        }
-        Rule::apply => {
-            let mut p = pair.into_inner();
-            let ident_p = p.next().ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "missing identifier in function application".to_string(),
-                }
-                .build()
-            })?;
-            let ident = ident_p.as_str();
-            let mut args: Vec<_> = p
-                .next()
-                .ok_or_else(|| {
-                    InvalidQuerySnafu {
-                        message: "missing arguments in function application".to_string(),
-                    }
-                    .build()
-                })?
-                .into_inner()
-                .map(|v| build_expr(v, param_pool))
-                .try_collect()?;
-
-            match ident {
-                "cond" => {
-                    if args.is_empty() {
-                        return Err(InvalidQuerySnafu {
-                            message: "'cond' cannot have empty body".to_string(),
-                        }
-                        .build()
-                        .into());
-                    }
-                    if args.len() & 1 == 1 {
-                        args.insert(
-                            args.len() - 1,
-                            Expr::Const {
-                                val: DataValue::Null,
-                                span: args
-                                    .last()
-                                    .ok_or_else(|| {
-                                        InvalidQuerySnafu {
-                                            message: "missing last argument for cond".to_string(),
-                                        }
-                                        .build()
-                                    })?
-                                    .span(),
-                            },
-                        )
-                    }
-                    let mut clauses = args
-                        .chunks(2)
-                        .map(|pair| (pair[0].clone(), pair[1].clone()))
-                        .collect_vec();
-                    if let Some((cond, _)) = clauses.last() {
-                        match cond {
-                            // NOTE: last clause already has true default, no null fallback needed
-                            Expr::Const {
-                                val: DataValue::Bool(true),
-                                ..
-                            } => {}
-                            _ => {
-                                clauses.push((
-                                    Expr::Const {
-                                        val: DataValue::from(true),
-                                        span,
-                                    },
-                                    Expr::Const {
-                                        val: DataValue::Null,
-                                        span,
-                                    },
-                                ));
-                            }
-                        }
-                    }
-                    Expr::Cond { clauses, span }
-                }
-                "if" => {
-                    if args.len() != 2 && args.len() != 3 {
-                        return Err(InvalidQuerySnafu {
-                            message: "wrong number of arguments to if: 2 or 3 required".to_string(),
-                        }
-                        .build()
-                        .into());
-                    }
-
-                    let mut clauses = vec![];
-                    let mut args = args.into_iter();
-                    let cond = args.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "missing condition in if expression".to_string(),
-                        }
-                        .build()
-                    })?;
-                    let then = args.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "missing then branch in if expression".to_string(),
-                        }
-                        .build()
-                    })?;
-                    clauses.push((cond, then));
-                    clauses.push((
-                        Expr::Const {
-                            val: DataValue::from(true),
-                            span,
-                        },
-                        args.next().unwrap_or(Expr::Const {
-                            val: DataValue::Null,
-                            span,
-                        }),
-                    ));
-                    Expr::Cond { clauses, span }
-                }
-                _ => match get_op(ident) {
-                    None => Expr::UnboundApply {
-                        op: ident.into(),
-                        args: args.into(),
-                        span,
-                    },
-                    Some(op) => {
-                        op.post_process_args(&mut args);
-
-                        if op.vararg {
-                            if op.min_arity > args.len() {
-                                return Err(InvalidQuerySnafu {
-                                    message: format!(
-                                        "Wrong number of arguments for function: need at least {} argument(s)",
-                                        op.min_arity
-                                    )
-                                }
-                                .build().into());
-                            }
-                        } else if op.min_arity != args.len() {
-                            return Err(InvalidQuerySnafu {
-                                message: format!(
-                                    "Wrong number of arguments for function: need exactly {} argument(s)",
-                                    op.min_arity
-                                )
-                            }
-                            .build().into());
-                        }
-                        Expr::Apply {
-                            op,
-                            args: args.into(),
-                            span,
-                        }
-                    }
-                },
-            }
-        }
-        Rule::grouping => build_expr(
-            pair.into_inner().next().ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "empty grouping expression".to_string(),
-                }
-                .build()
-            })?,
-            param_pool,
-        )?,
-        r => {
-            return Err(InvalidQuerySnafu {
-                message: format!(
-                    "unexpected rule {:?} in parser - grammar mismatch, please file a bug",
-                    r
-                ),
-            }
-            .build()
-            .into());
-        }
-    })
-}
-
-pub(crate) fn parse_int(s: &str, radix: u32) -> i64 {
-    // WHY: get(2..) skips "0x"/"0o"/"0b" prefix; grammar guarantees valid integer format.
-    i64::from_str_radix(&s.get(2..).unwrap_or("").replace('_', ""), radix)
-        .unwrap_or_else(|_| unreachable!("INVARIANT: grammar guarantees valid integer format after prefix"))
-}
-
-pub(crate) fn parse_string(pair: Pair<'_>) -> Result<CompactString> {
-    match pair.as_rule() {
-        Rule::quoted_string => Ok(parse_quoted_string(pair)?),
-        Rule::s_quoted_string => Ok(parse_s_quoted_string(pair)?),
-        Rule::raw_string => Ok(parse_raw_string(pair)?),
-        Rule::ident => Ok(CompactString::from(pair.as_str())),
-        r => {
-            return Err(InvalidQuerySnafu {
-                message: format!(
-                    "unexpected rule {:?} in parser - grammar mismatch, please file a bug",
-                    r
-                ),
-            }
-            .build()
-            .into());
-        }
-    }
-}
-
-fn parse_quoted_string(pair: Pair<'_>) -> Result<CompactString> {
-    let pairs = pair
-        .into_inner()
-        .next()
-        .ok_or_else(|| {
-            InvalidQuerySnafu {
-                message: "empty quoted string".to_string(),
-            }
-            .build()
-        })?
-        .into_inner();
-    let mut ret = CompactString::default();
-    for pair in pairs {
-        let s = pair.as_str();
-        match s {
-            r#"\""# => ret.push('"'),
-            r"\\" => ret.push('\\'),
-            r"\/" => ret.push('/'),
-            r"\b" => ret.push('\x08'),
-            r"\f" => ret.push('\x0c'),
-            r"\n" => ret.push('\n'),
-            r"\r" => ret.push('\r'),
-            r"\t" => ret.push('\t'),
-            s if s.starts_with(r"\u") => {
-                #[expect(clippy::cast_possible_truncation, reason = "value fits u32")]
-                let code = parse_int(s, 16) as u32;
-                let ch = char::from_u32(code).ok_or_else(|| {
-                    InvalidQuerySnafu {
-                        message: format!("invalid UTF8 code {code}"),
-                    }
-                    .build()
-                })?;
-                ret.push(ch);
-            }
-            s if s.starts_with('\\') => {
-                return Err(InvalidQuerySnafu {
-                    message: format!("invalid escape sequence {s}"),
-                }
-                .build()
-                .into());
-            }
-            s => ret.push_str(s),
-        }
-    }
-    Ok(ret)
-}
-
-fn parse_s_quoted_string(pair: Pair<'_>) -> Result<CompactString> {
-    let pairs = pair
-        .into_inner()
-        .next()
-        .ok_or_else(|| {
-            InvalidQuerySnafu {
-                message: "empty single-quoted string".to_string(),
-            }
-            .build()
-        })?
-        .into_inner();
-    let mut ret = CompactString::default();
-    for pair in pairs {
-        let s = pair.as_str();
-        match s {
-            r#"\'"# => ret.push('\''),
-            r"\\" => ret.push('\\'),
-            r"\/" => ret.push('/'),
-            r"\b" => ret.push('\x08'),
-            r"\f" => ret.push('\x0c'),
-            r"\n" => ret.push('\n'),
-            r"\r" => ret.push('\r'),
-            r"\t" => ret.push('\t'),
-            s if s.starts_with(r"\u") => {
-                #[expect(clippy::cast_possible_truncation, reason = "value fits u32")]
-                let code = parse_int(s, 16) as u32;
-                let ch = char::from_u32(code).ok_or_else(|| {
-                    InvalidQuerySnafu {
-                        message: format!("invalid UTF8 code {code}"),
-                    }
-                    .build()
-                })?;
-                ret.push(ch);
-            }
-            s if s.starts_with('\\') => {
-                return Err(InvalidQuerySnafu {
-                    message: format!("invalid escape sequence {s}"),
-                }
-                .build()
-                .into());
-            }
-            s => ret.push_str(s),
-        }
-    }
-    Ok(ret)
-}
-
-fn parse_raw_string(pair: Pair<'_>) -> Result<CompactString> {
-    Ok(CompactString::from(
-        pair.into_inner()
-            .next()
-            .ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "empty raw string".to_string(),
-                }
-                .build()
-            })?
-            .as_str(),
-    ))
-}
