+++ crates/krites/src/parse/sys/parse.rs
@@ -1,4 +1,7 @@
-//! `parse_sys` implementation.
+//! System command dispatch: routes each `::` command to the appropriate parser.
+//!
+//! Index creation for FTS, HNSW, and LSH is delegated to [`super::index`] to
+//! keep this file focused on top-level dispatch.
 #![expect(
     clippy::pedantic,
     clippy::result_large_err,
@@ -8,23 +11,32 @@
 use std::collections::BTreeMap;
 use std::sync::Arc;
 
-use compact_str::CompactString;
 use itertools::Itertools;
 
-use crate::data::relation::VecElementType;
 use crate::data::symb::Symbol;
 use crate::data::value::{DataValue, ValidityTs};
 use crate::error::InternalResult as Result;
-use crate::fts::TokenizerConfig;
 use crate::parse::error::InvalidQuerySnafu;
 use crate::parse::expr::{build_expr, parse_string};
 use crate::parse::query::parse_query;
 use crate::parse::{ExtractSpan, Pairs, Rule};
 use crate::runtime::relation::AccessLevel;
-use crate::{Expr, FixedRule};
+use crate::FixedRule;
 
-use super::{FtsIndexConfig, HnswDistance, HnswIndexConfig, MinHashLshConfig, SysOp};
+use super::SysOp;
+use super::index::{
+    parse_fts_index_create, parse_hnsw_index_create, parse_index_drop, parse_lsh_index_create,
+};
 
+/// Parse the inner pairs of a `sys_script` into a [`SysOp`].
+///
+/// Each system command is a `::keyword` form. This function dispatches on
+/// the pest rule to the appropriate handler.
+///
+/// # Errors
+///
+/// Returns an error if the command syntax is invalid or an option value
+/// cannot be evaluated.
 pub(crate) fn parse_sys(
     mut src: Pairs<'_>,
     param_pool: &BTreeMap<String, DataValue>,
@@ -33,962 +45,38 @@ pub(crate) fn parse_sys(
 ) -> Result<SysOp> {
     let inner = src.next().ok_or_else(|| {
         InvalidQuerySnafu {
-            message: "expected element".to_string(),
+            message: "expected system command".to_string(),
         }
         .build()
     })?;
     Ok(match inner.as_rule() {
         Rule::compact_op => SysOp::Compact,
         Rule::running_op => SysOp::ListRunning,
-        Rule::kill_op => {
-            let i_expr = inner.into_inner().next().ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "expected element".to_string(),
-                }
-                .build()
-            })?;
-            let i_val = build_expr(i_expr, param_pool)?;
-            let i_val = i_val.eval_to_const()?;
-            let i_val = i_val.get_int().ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "Process ID must be an integer".to_string(),
-                }
-                .build()
-            })?;
-            let pid = u64::try_from(i_val).map_err(|_e| {
-                InvalidQuerySnafu {
-                    message: "Process ID must be a non-negative integer".to_string(),
-                }
-                .build()
-            })?;
-            SysOp::KillRunning(pid)
-        }
-        Rule::explain_op => {
-            let prog = parse_query(
-                inner
-                    .into_inner()
-                    .next()
-                    .ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?
-                    .into_inner(),
-                param_pool,
-                algorithms,
-                cur_vld,
-            )?;
-            SysOp::Explain(Box::new(prog))
-        }
-        Rule::describe_relation_op => {
-            let mut inner = inner.into_inner();
-            let rels_p = inner.next().ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "expected element".to_string(),
-                }
-                .build()
-            })?;
-            let rel = Symbol::new(rels_p.as_str(), rels_p.extract_span());
-            let description = match inner.next() {
-                None => Default::default(),
-                Some(desc_p) => parse_string(desc_p)?,
-            };
-            SysOp::DescribeRelation(rel, description)
-        }
+        Rule::kill_op => parse_kill_op(inner, param_pool)?,
+        Rule::explain_op => parse_explain_op(inner, param_pool, algorithms, cur_vld)?,
+        Rule::describe_relation_op => parse_describe_op(inner)?,
         Rule::list_relations_op => SysOp::ListRelations,
         Rule::remove_relations_op => {
             let rel = inner
                 .into_inner()
                 .map(|rels_p| Symbol::new(rels_p.as_str(), rels_p.extract_span()))
                 .collect_vec();
-
             SysOp::RemoveRelation(rel)
         }
-        Rule::list_columns_op => {
-            let rels_p = inner.into_inner().next().ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "expected element".to_string(),
-                }
-                .build()
-            })?;
-            let rel = Symbol::new(rels_p.as_str(), rels_p.extract_span());
-            SysOp::ListColumns(rel)
-        }
-        Rule::list_indices_op => {
-            let rels_p = inner.into_inner().next().ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "expected element".to_string(),
-                }
-                .build()
-            })?;
-            let rel = Symbol::new(rels_p.as_str(), rels_p.extract_span());
-            SysOp::ListIndices(rel)
-        }
-        Rule::rename_relations_op => {
-            let mut rename_pairs = vec![];
-            for pair in inner.into_inner() {
-                let mut src = pair.into_inner();
-                let rels_p = src.next().ok_or_else(|| {
-                    InvalidQuerySnafu {
-                        message: "expected element in rename relations".to_string(),
-                    }
-                    .build()
-                })?;
-                let rel = Symbol::new(rels_p.as_str(), rels_p.extract_span());
-                let rels_p = src.next().ok_or_else(|| {
-                    InvalidQuerySnafu {
-                        message: "expected element in rename relations".to_string(),
-                    }
-                    .build()
-                })?;
-                let new_rel = Symbol::new(rels_p.as_str(), rels_p.extract_span());
-                rename_pairs.push((rel, new_rel));
-            }
-            SysOp::RenameRelation(rename_pairs)
-        }
-        Rule::access_level_op => {
-            let mut ps = inner.into_inner();
-            let access_level = match ps
-                .next()
-                .ok_or_else(|| {
-                    InvalidQuerySnafu {
-                        message: "expected element".to_string(),
-                    }
-                    .build()
-                })?
-                .as_str()
-            {
-                "normal" => AccessLevel::Normal,
-                "protected" => AccessLevel::Protected,
-                "read_only" => AccessLevel::ReadOnly,
-                "hidden" => AccessLevel::Hidden,
-                s => {
-                    return Err(InvalidQuerySnafu {
-                        message: format!("unexpected access level: {}", s),
-                    }
-                    .build()
-                    .into());
-                }
-            };
-            let mut rels = vec![];
-            for rel_p in ps {
-                let rel = Symbol::new(rel_p.as_str(), rel_p.extract_span());
-                rels.push(rel)
-            }
-            SysOp::SetAccessLevel(rels, access_level)
-        }
+        Rule::list_columns_op => parse_single_relation_op(inner, SysOp::ListColumns)?,
+        Rule::list_indices_op => parse_single_relation_op(inner, SysOp::ListIndices)?,
+        Rule::rename_relations_op => parse_rename_op(inner)?,
+        Rule::access_level_op => parse_access_level_op(inner)?,
         Rule::trigger_relation_show_op => {
-            let rels_p = inner.into_inner().next().ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "expected element".to_string(),
-                }
-                .build()
-            })?;
-            let rel = Symbol::new(rels_p.as_str(), rels_p.extract_span());
-            SysOp::ShowTrigger(rel)
+            parse_single_relation_op(inner, SysOp::ShowTrigger)?
         }
         Rule::trigger_relation_op => {
-            let mut src = inner.into_inner();
-            let rels_p = src.next().ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "expected element".to_string(),
-                }
-                .build()
-            })?;
-            let rel = Symbol::new(rels_p.as_str(), rels_p.extract_span());
-            let mut puts = vec![];
-            let mut rms = vec![];
-            let mut replaces = vec![];
-            for clause in src {
-                let mut clause_inner = clause.into_inner();
-                let op = clause_inner.next().ok_or_else(|| {
-                    InvalidQuerySnafu {
-                        message: "expected element".to_string(),
-                    }
-                    .build()
-                })?;
-                let script = clause_inner.next().ok_or_else(|| {
-                    InvalidQuerySnafu {
-                        message: "expected element".to_string(),
-                    }
-                    .build()
-                })?;
-                let script_str = script.as_str();
-                parse_query(
-                    script.into_inner(),
-                    &Default::default(),
-                    algorithms,
-                    cur_vld,
-                )?;
-                match op.as_rule() {
-                    Rule::trigger_put => puts.push(script_str.to_string()),
-                    Rule::trigger_rm => rms.push(script_str.to_string()),
-                    Rule::trigger_replace => replaces.push(script_str.to_string()),
-                    r => {
-                        return Err(InvalidQuerySnafu {
-                            message: format!("unexpected rule {:?} in system parser", r),
-                        }
-                        .build()
-                        .into());
-                    }
-                }
-            }
-            SysOp::SetTriggers(rel, puts, rms, replaces)
+            parse_trigger_op(inner, param_pool, algorithms, cur_vld)?
         }
-        Rule::lsh_idx_op => {
-            let inner = inner.into_inner().next().ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "expected element".to_string(),
-                }
-                .build()
-            })?;
-            match inner.as_rule() {
-                Rule::index_create_adv => {
-                    let mut inner = inner.into_inner();
-                    let rel = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    let name = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    let mut filters = vec![];
-                    let mut tokenizer = TokenizerConfig {
-                        name: Default::default(),
-                        args: Default::default(),
-                    };
-                    let mut extractor = "".to_string();
-                    let mut extract_filter = "".to_string();
-                    let mut n_gram = 1;
-                    let mut n_perm = 200;
-                    let mut target_threshold = 0.9;
-                    let mut false_positive_weight = 1.0;
-                    let mut false_negative_weight = 1.0;
-                    for opt_pair in inner {
-                        let mut opt_inner = opt_pair.into_inner();
-                        let opt_name = opt_inner.next().ok_or_else(|| {
-                            InvalidQuerySnafu {
-                                message: "expected element".to_string(),
-                            }
-                            .build()
-                        })?;
-                        let opt_val = opt_inner.next().ok_or_else(|| {
-                            InvalidQuerySnafu {
-                                message: "expected element".to_string(),
-                            }
-                            .build()
-                        })?;
-                        match opt_name.as_str() {
-                            "false_positive_weight" => {
-                                let mut expr = build_expr(opt_val, param_pool)?;
-                                expr.partial_eval()?;
-                                let v = expr.eval_to_const()?;
-                                false_positive_weight = v.get_float().ok_or_else(|| {
-                                    InvalidQuerySnafu {
-                                        message: "false_positive_weight must be a float"
-                                            .to_string(),
-                                    }
-                                    .build()
-                                })?;
-                            }
-                            "false_negative_weight" => {
-                                let mut expr = build_expr(opt_val, param_pool)?;
-                                expr.partial_eval()?;
-                                let v = expr.eval_to_const()?;
-                                false_negative_weight = v.get_float().ok_or_else(|| {
-                                    InvalidQuerySnafu {
-                                        message: "false_negative_weight must be a float"
-                                            .to_string(),
-                                    }
-                                    .build()
-                                })?;
-                            }
-                            "n_gram" => {
-                                let mut expr = build_expr(opt_val, param_pool)?;
-                                expr.partial_eval()?;
-                                let v = expr.eval_to_const()?;
-                                let v_int = v.get_int().ok_or_else(|| {
-                                    InvalidQuerySnafu {
-                                        message: "n_gram must be an integer".to_string(),
-                                    }
-                                    .build()
-                                })?;
-                                n_gram = usize::try_from(v_int).map_err(|_e| {
-                                    InvalidQuerySnafu {
-                                        message: "n_gram must be a non-negative integer"
-                                            .to_string(),
-                                    }
-                                    .build()
-                                })?;
-                            }
-                            "n_perm" => {
-                                let mut expr = build_expr(opt_val, param_pool)?;
-                                expr.partial_eval()?;
-                                let v = expr.eval_to_const()?;
-                                let v_int = v.get_int().ok_or_else(|| {
-                                    InvalidQuerySnafu {
-                                        message: "n_perm must be an integer".to_string(),
-                                    }
-                                    .build()
-                                })?;
-                                n_perm = usize::try_from(v_int).map_err(|_e| {
-                                    InvalidQuerySnafu {
-                                        message: "n_perm must be a non-negative integer"
-                                            .to_string(),
-                                    }
-                                    .build()
-                                })?;
-                            }
-                            "target_threshold" => {
-                                let mut expr = build_expr(opt_val, param_pool)?;
-                                expr.partial_eval()?;
-                                let v = expr.eval_to_const()?;
-                                target_threshold = v.get_float().ok_or_else(|| {
-                                    InvalidQuerySnafu {
-                                        message: "target_threshold must be a float".to_string(),
-                                    }
-                                    .build()
-                                })?;
-                            }
-                            "extractor" => {
-                                let mut ex = build_expr(opt_val, param_pool)?;
-                                ex.partial_eval()?;
-                                extractor = ex.to_string();
-                            }
-                            "extract_filter" => {
-                                let mut ex = build_expr(opt_val, param_pool)?;
-                                ex.partial_eval()?;
-                                extract_filter = ex.to_string();
-                            }
-                            "tokenizer" => {
-                                let mut expr = build_expr(opt_val, param_pool)?;
-                                expr.partial_eval()?;
-                                match expr {
-                                    Expr::UnboundApply { op, args, .. } => {
-                                        let mut targs = vec![];
-                                        for arg in args.iter() {
-                                            let v = arg.clone().eval_to_const()?;
-                                            targs.push(v);
-                                        }
-                                        tokenizer.name = op;
-                                        tokenizer.args = targs;
-                                    }
-                                    Expr::Binding { var, .. } => {
-                                        tokenizer.name = var.name;
-                                        tokenizer.args = vec![];
-                                    }
-                                    _ => return Err(InvalidQuerySnafu {
-                                        message: "Tokenizer must be a symbol or a call for an existing tokenizer".to_string()
-                                    }.build().into()),
-                                }
-                            }
-                            "filters" => {
-                                let mut expr = build_expr(opt_val, param_pool)?;
-                                expr.partial_eval()?;
-                                match expr {
-                                    Expr::Apply { op, args, .. } => {
-                                        if op.name != "OP_LIST" {
-                                            return Err(InvalidQuerySnafu {
-                                                message: "Filters must be a list of filters"
-                                                    .to_string(),
-                                            }
-                                            .build()
-                                            .into());
-                                        }
-                                        for arg in args.iter() {
-                                            match arg {
-                                                Expr::UnboundApply { op, args, .. } => {
-                                                    let mut targs = vec![];
-                                                    for arg in args.iter() {
-                                                        let v = arg.clone().eval_to_const()?;
-                                                        targs.push(v);
-                                                    }
-                                                    filters.push(TokenizerConfig {
-                                                        name: op.clone(),
-                                                        args: targs,
-                                                    })
-                                                }
-                                                Expr::Binding { var, .. } => {
-                                                    filters.push(TokenizerConfig {
-                                                        name: var.name.clone(),
-                                                        args: vec![],
-                                                    })
-                                                }
-                                                _ => return Err(InvalidQuerySnafu {
-                                                    message: "Tokenizer must be a symbol or a call for an existing tokenizer".to_string()
-                                                }
-                                                .build().into()),
-                                            }
-                                        }
-                                    }
-                                    _ => {
-                                        return Err(InvalidQuerySnafu {
-                                            message: "Filters must be a list of filters"
-                                                .to_string(),
-                                        }
-                                        .build()
-                                        .into());
-                                    }
-                                }
-                            }
-                            s => {
-                                return Err(InvalidQuerySnafu {
-                                    message: format!("Unknown option {s} for LSH index"),
-                                }
-                                .build()
-                                .into());
-                            }
-                        }
-                    }
-                    if false_positive_weight <= 0. {
-                        return Err(InvalidQuerySnafu {
-                            message: "false_positive_weight must be positive".to_string(),
-                        }
-                        .build()
-                        .into());
-                    }
-                    if false_negative_weight <= 0. {
-                        return Err(InvalidQuerySnafu {
-                            message: "false_negative_weight must be positive".to_string(),
-                        }
-                        .build()
-                        .into());
-                    }
-                    if n_gram == 0 {
-                        return Err(InvalidQuerySnafu {
-                            message: "n_gram must be positive".to_string(),
-                        }
-                        .build()
-                        .into());
-                    }
-                    if n_perm == 0 {
-                        return Err(InvalidQuerySnafu {
-                            message: "n_perm must be positive".to_string(),
-                        }
-                        .build()
-                        .into());
-                    }
-                    if target_threshold <= 0. || target_threshold >= 1. {
-                        return Err(InvalidQuerySnafu {
-                            message: "target_threshold must be between 0 and 1".to_string(),
-                        }
-                        .build()
-                        .into());
-                    }
-                    let total_weights = false_positive_weight + false_negative_weight;
-                    false_positive_weight /= total_weights;
-                    false_negative_weight /= total_weights;
-
-                    if !extract_filter.is_empty() {
-                        extractor = format!("if({extract_filter}, {extractor})");
-                    }
-
-                    let config = MinHashLshConfig {
-                        base_relation: CompactString::from(rel.as_str()),
-                        index_name: CompactString::from(name.as_str()),
-                        extractor,
-                        tokenizer,
-                        filters,
-                        n_gram,
-                        n_perm,
-                        false_positive_weight: false_positive_weight.into(),
-                        false_negative_weight: false_negative_weight.into(),
-                        target_threshold: target_threshold.into(),
-                    };
-                    SysOp::CreateMinHashLshIndex(config)
-                }
-                Rule::index_drop => {
-                    let mut inner = inner.into_inner();
-                    let rel = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    let name = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    SysOp::RemoveIndex(
-                        Symbol::new(rel.as_str(), rel.extract_span()),
-                        Symbol::new(name.as_str(), name.extract_span()),
-                    )
-                }
-                r => {
-                    return Err(InvalidQuerySnafu {
-                        message: format!("unexpected rule {:?} in system parser", r),
-                    }
-                    .build()
-                    .into());
-                }
-            }
-        }
-        Rule::fts_idx_op => {
-            let inner = inner.into_inner().next().ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "expected element".to_string(),
-                }
-                .build()
-            })?;
-            match inner.as_rule() {
-                Rule::index_create_adv => {
-                    let mut inner = inner.into_inner();
-                    let rel = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    let name = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    let mut filters = vec![];
-                    let mut tokenizer = TokenizerConfig {
-                        name: Default::default(),
-                        args: Default::default(),
-                    };
-                    let mut extractor = "".to_string();
-                    let mut extract_filter = "".to_string();
-                    for opt_pair in inner {
-                        let mut opt_inner = opt_pair.into_inner();
-                        let opt_name = opt_inner.next().ok_or_else(|| {
-                            InvalidQuerySnafu {
-                                message: "expected element".to_string(),
-                            }
-                            .build()
-                        })?;
-                        let opt_val = opt_inner.next().ok_or_else(|| {
-                            InvalidQuerySnafu {
-                                message: "expected element".to_string(),
-                            }
-                            .build()
-                        })?;
-                        match opt_name.as_str() {
-                            "extractor" => {
-                                let mut ex = build_expr(opt_val, param_pool)?;
-                                ex.partial_eval()?;
-                                extractor = ex.to_string();
-                            }
-                            "extract_filter" => {
-                                let mut ex = build_expr(opt_val, param_pool)?;
-                                ex.partial_eval()?;
-                                extract_filter = ex.to_string();
-                            }
-                            "tokenizer" => {
-                                let mut expr = build_expr(opt_val, param_pool)?;
-                                expr.partial_eval()?;
-                                match expr {
-                                    Expr::UnboundApply { op, args, .. } => {
-                                        let mut targs = vec![];
-                                        for arg in args.iter() {
-                                            let v = arg.clone().eval_to_const()?;
-                                            targs.push(v);
-                                        }
-                                        tokenizer.name = op;
-                                        tokenizer.args = targs;
-                                    }
-                                    Expr::Binding { var, .. } => {
-                                        tokenizer.name = var.name;
-                                        tokenizer.args = vec![];
-                                    }
-                                    _ => return Err(InvalidQuerySnafu {
-                                        message: "Tokenizer must be a symbol or a call for an existing tokenizer".to_string()
-                                    }
-                                    .build().into()),
-                                }
-                            }
-                            "filters" => {
-                                let mut expr = build_expr(opt_val, param_pool)?;
-                                expr.partial_eval()?;
-                                match expr {
-                                    Expr::Apply { op, args, .. } => {
-                                        if op.name != "OP_LIST" {
-                                            return Err(InvalidQuerySnafu {
-                                                message: "Filters must be a list of filters"
-                                                    .to_string(),
-                                            }
-                                            .build()
-                                            .into());
-                                        }
-                                        for arg in args.iter() {
-                                            match arg {
-                                                Expr::UnboundApply { op, args, .. } => {
-                                                    let mut targs = vec![];
-                                                    for arg in args.iter() {
-                                                        let v = arg.clone().eval_to_const()?;
-                                                        targs.push(v);
-                                                    }
-                                                    filters.push(TokenizerConfig {
-                                                        name: op.clone(),
-                                                        args: targs,
-                                                    })
-                                                }
-                                                Expr::Binding { var, .. } => {
-                                                    filters.push(TokenizerConfig {
-                                                        name: var.name.clone(),
-                                                        args: vec![],
-                                                    })
-                                                }
-                                                _ => return Err(InvalidQuerySnafu {
-                                                    message: "Tokenizer must be a symbol or a call for an existing tokenizer".to_string()
-                                                }
-                                                .build().into()),
-                                            }
-                                        }
-                                    }
-                                    _ => {
-                                        return Err(InvalidQuerySnafu {
-                                            message: "Filters must be a list of filters"
-                                                .to_string(),
-                                        }
-                                        .build()
-                                        .into());
-                                    }
-                                }
-                            }
-                            s => {
-                                return Err(InvalidQuerySnafu {
-                                    message: format!("Unknown option {s} for FTS index"),
-                                }
-                                .build()
-                                .into());
-                            }
-                        }
-                    }
-                    if !extract_filter.is_empty() {
-                        extractor = format!("if({extract_filter}, {extractor})");
-                    }
-                    let config = FtsIndexConfig {
-                        base_relation: CompactString::from(rel.as_str()),
-                        index_name: CompactString::from(name.as_str()),
-                        extractor,
-                        tokenizer,
-                        filters,
-                    };
-                    SysOp::CreateFtsIndex(config)
-                }
-                Rule::index_drop => {
-                    let mut inner = inner.into_inner();
-                    let rel = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    let name = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    SysOp::RemoveIndex(
-                        Symbol::new(rel.as_str(), rel.extract_span()),
-                        Symbol::new(name.as_str(), name.extract_span()),
-                    )
-                }
-                r => {
-                    return Err(InvalidQuerySnafu {
-                        message: format!("unexpected rule {:?} in system parser", r),
-                    }
-                    .build()
-                    .into());
-                }
-            }
-        }
-        Rule::vec_idx_op => {
-            let inner = inner.into_inner().next().ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "expected element".to_string(),
-                }
-                .build()
-            })?;
-            match inner.as_rule() {
-                Rule::index_create_adv => {
-                    let mut inner = inner.into_inner();
-                    let rel = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    let name = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    let mut vec_dim = 0;
-                    let mut dtype = VecElementType::F32;
-                    let mut vec_fields = vec![];
-                    let mut distance = HnswDistance::L2;
-                    let mut ef_construction = 0;
-                    let mut m_neighbours = 0;
-                    let mut index_filter = None;
-                    let mut extend_candidates = false;
-                    let mut keep_pruned_connections = false;
-
-                    for opt_pair in inner {
-                        let mut opt_inner = opt_pair.into_inner();
-                        let opt_name = opt_inner.next().ok_or_else(|| {
-                            InvalidQuerySnafu {
-                                message: "expected element".to_string(),
-                            }
-                            .build()
-                        })?;
-                        let opt_val = opt_inner.next().ok_or_else(|| {
-                            InvalidQuerySnafu {
-                                message: "expected element".to_string(),
-                            }
-                            .build()
-                        })?;
-                        let opt_val_str = opt_val.as_str();
-                        match opt_name.as_str() {
-                            "dim" => {
-                                let v = build_expr(opt_val, param_pool)?
-                                    .eval_to_const()?
-                                    .get_int()
-                                    .ok_or_else(|| {
-                                        InvalidQuerySnafu {
-                                            message: format!("Invalid vec_dim: {opt_val_str}"),
-                                        }
-                                        .build()
-                                    })?;
-                                if v <= 0 {
-                                    return Err(InvalidQuerySnafu {
-                                        message: format!("Invalid vec_dim: {v}"),
-                                    }
-                                    .build()
-                                    .into());
-                                }
-                                // INVARIANT: range-checked > 0 above.
-                                vec_dim = usize::try_from(v).unwrap_or(usize::MAX);
-                            }
-                            "ef_construction" | "ef" => {
-                                let v = build_expr(opt_val, param_pool)?
-                                    .eval_to_const()?
-                                    .get_int()
-                                    .ok_or_else(|| {
-                                        InvalidQuerySnafu {
-                                            message: format!(
-                                                "Invalid ef_construction: {opt_val_str}"
-                                            ),
-                                        }
-                                        .build()
-                                    })?;
-                                if v <= 0 {
-                                    return Err(InvalidQuerySnafu {
-                                        message: format!("Invalid ef_construction: {v}"),
-                                    }
-                                    .build()
-                                    .into());
-                                }
-                                // INVARIANT: range-checked > 0 above.
-                                ef_construction = usize::try_from(v).unwrap_or(usize::MAX);
-                            }
-                            "m_neighbours" | "m" => {
-                                let v = build_expr(opt_val, param_pool)?
-                                    .eval_to_const()?
-                                    .get_int()
-                                    .ok_or_else(|| {
-                                        InvalidQuerySnafu {
-                                            message: format!("Invalid m_neighbours: {opt_val_str}"),
-                                        }
-                                        .build()
-                                    })?;
-                                if v <= 0 {
-                                    return Err(InvalidQuerySnafu {
-                                        message: format!("Invalid m_neighbours: {v}"),
-                                    }
-                                    .build()
-                                    .into());
-                                }
-                                // INVARIANT: range-checked > 0 above.
-                                m_neighbours = usize::try_from(v).unwrap_or(usize::MAX);
-                            }
-                            "dtype" => {
-                                dtype = match opt_val.as_str() {
-                                    "F32" | "Float" => VecElementType::F32,
-                                    "F64" | "Double" => VecElementType::F64,
-                                    s => {
-                                        return Err(InvalidQuerySnafu {
-                                            message: format!("Invalid dtype: {s}"),
-                                        }
-                                        .build()
-                                        .into());
-                                    }
-                                }
-                            }
-                            "fields" => {
-                                let fields = build_expr(opt_val, &Default::default())?;
-                                vec_fields = fields.to_var_list()?;
-                            }
-                            "distance" | "dist" => {
-                                distance = match opt_val.as_str().trim() {
-                                    "L2" => HnswDistance::L2,
-                                    "IP" => HnswDistance::InnerProduct,
-                                    "Cosine" => HnswDistance::Cosine,
-                                    s => {
-                                        return Err(InvalidQuerySnafu {
-                                            message: format!("Invalid distance: {s}"),
-                                        }
-                                        .build()
-                                        .into());
-                                    }
-                                }
-                            }
-                            "filter" => {
-                                index_filter = Some(opt_val.as_str().to_string());
-                            }
-                            "extend_candidates" => {
-                                extend_candidates = opt_val.as_str().trim() == "true";
-                            }
-                            "keep_pruned_connections" => {
-                                keep_pruned_connections = opt_val.as_str().trim() == "true";
-                            }
-                            s => {
-                                return Err(InvalidQuerySnafu {
-                                    message: format!("Invalid option: {s}"),
-                                }
-                                .build()
-                                .into());
-                            }
-                        }
-                    }
-                    if ef_construction == 0 {
-                        return Err(InvalidQuerySnafu {
-                            message: "ef_construction must be set".to_string(),
-                        }
-                        .build()
-                        .into());
-                    }
-                    if m_neighbours == 0 {
-                        return Err(InvalidQuerySnafu {
-                            message: "m_neighbours must be set".to_string(),
-                        }
-                        .build()
-                        .into());
-                    }
-                    SysOp::CreateVectorIndex(HnswIndexConfig {
-                        base_relation: CompactString::from(rel.as_str()),
-                        index_name: CompactString::from(name.as_str()),
-                        vec_dim,
-                        dtype,
-                        vec_fields,
-                        distance,
-                        ef_construction,
-                        m_neighbours,
-                        index_filter,
-                        extend_candidates,
-                        keep_pruned_connections,
-                    })
-                }
-                Rule::index_drop => {
-                    let mut inner = inner.into_inner();
-                    let rel = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    let name = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    SysOp::RemoveIndex(
-                        Symbol::new(rel.as_str(), rel.extract_span()),
-                        Symbol::new(name.as_str(), name.extract_span()),
-                    )
-                }
-                r => {
-                    return Err(InvalidQuerySnafu {
-                        message: format!("unexpected rule {:?} in system parser", r),
-                    }
-                    .build()
-                    .into());
-                }
-            }
-        }
-        Rule::index_op => {
-            let inner = inner.into_inner().next().ok_or_else(|| {
-                InvalidQuerySnafu {
-                    message: "expected element".to_string(),
-                }
-                .build()
-            })?;
-            match inner.as_rule() {
-                Rule::index_create => {
-                    let _span = inner.extract_span();
-                    let mut inner = inner.into_inner();
-                    let rel = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    let name = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    let cols = inner
-                        .map(|p| Symbol::new(p.as_str(), p.extract_span()))
-                        .collect_vec();
-
-                    if cols.is_empty() {
-                        return Err(InvalidQuerySnafu {
-                            message: "index must have at least one column specified".to_string(),
-                        }
-                        .build()
-                        .into());
-                    }
-                    SysOp::CreateIndex(
-                        Symbol::new(rel.as_str(), rel.extract_span()),
-                        Symbol::new(name.as_str(), name.extract_span()),
-                        cols,
-                    )
-                }
-                Rule::index_drop => {
-                    let mut inner = inner.into_inner();
-                    let rel = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    let name = inner.next().ok_or_else(|| {
-                        InvalidQuerySnafu {
-                            message: "expected element".to_string(),
-                        }
-                        .build()
-                    })?;
-                    SysOp::RemoveIndex(
-                        Symbol::new(rel.as_str(), rel.extract_span()),
-                        Symbol::new(name.as_str(), name.extract_span()),
-                    )
-                }
-                r => {
-                    return Err(InvalidQuerySnafu {
-                        message: format!("unexpected rule {:?} in system parser", r),
-                    }
-                    .build()
-                    .into());
-                }
-            }
-        }
+        Rule::lsh_idx_op => parse_adv_index_op(inner, param_pool, parse_lsh_index_create)?,
+        Rule::fts_idx_op => parse_adv_index_op(inner, param_pool, parse_fts_index_create)?,
+        Rule::vec_idx_op => parse_adv_index_op(inner, param_pool, parse_hnsw_index_create)?,
+        Rule::index_op => parse_standard_index_op(inner)?,
         Rule::list_fixed_rules => SysOp::ListFixedRules,
         r => {
             return Err(InvalidQuerySnafu {
@@ -998,4 +86,277 @@ pub(crate) fn parse_sys(
             .into());
         }
     })
+}
+
+/// Parse `::kill <id>`.
+fn parse_kill_op(
+    inner: crate::parse::Pair<'_>,
+    param_pool: &BTreeMap<String, DataValue>,
+) -> Result<SysOp> {
+    let i_expr = inner.into_inner().next().ok_or_else(|| {
+        InvalidQuerySnafu {
+            message: "expected process ID".to_string(),
+        }
+        .build()
+    })?;
+    let i_val = build_expr(i_expr, param_pool)?.eval_to_const()?;
+    let i_val = i_val.get_int().ok_or_else(|| {
+        InvalidQuerySnafu {
+            message: "Process ID must be an integer".to_string(),
+        }
+        .build()
+    })?;
+    let pid = u64::try_from(i_val).map_err(|_e| {
+        InvalidQuerySnafu {
+            message: "Process ID must be a non-negative integer".to_string(),
+        }
+        .build()
+    })?;
+    Ok(SysOp::KillRunning(pid))
+}
+
+/// Parse `::explain { query }`.
+fn parse_explain_op(
+    inner: crate::parse::Pair<'_>,
+    param_pool: &BTreeMap<String, DataValue>,
+    algorithms: &BTreeMap<String, Arc<Box<dyn FixedRule>>>,
+    cur_vld: ValidityTs,
+) -> Result<SysOp> {
+    let prog = parse_query(
+        inner
+            .into_inner()
+            .next()
+            .ok_or_else(|| {
+                InvalidQuerySnafu {
+                    message: "expected query in explain".to_string(),
+                }
+                .build()
+            })?
+            .into_inner(),
+        param_pool,
+        algorithms,
+        cur_vld,
+    )?;
+    Ok(SysOp::Explain(Box::new(prog)))
+}
+
+/// Parse `::describe <relation> [description]`.
+fn parse_describe_op(inner: crate::parse::Pair<'_>) -> Result<SysOp> {
+    let mut inner = inner.into_inner();
+    let rels_p = inner.next().ok_or_else(|| {
+        InvalidQuerySnafu {
+            message: "expected relation name".to_string(),
+        }
+        .build()
+    })?;
+    let rel = Symbol::new(rels_p.as_str(), rels_p.extract_span());
+    let description = match inner.next() {
+        None => Default::default(),
+        Some(desc_p) => parse_string(desc_p)?,
+    };
+    Ok(SysOp::DescribeRelation(rel, description))
+}
+
+/// Parse a system command that takes a single relation argument and wraps
+/// it in a `SysOp` variant (e.g., `ListColumns`, `ListIndices`, `ShowTrigger`).
+fn parse_single_relation_op(
+    inner: crate::parse::Pair<'_>,
+    constructor: fn(Symbol) -> SysOp,
+) -> Result<SysOp> {
+    let rels_p = inner.into_inner().next().ok_or_else(|| {
+        InvalidQuerySnafu {
+            message: "expected relation name".to_string(),
+        }
+        .build()
+    })?;
+    Ok(constructor(Symbol::new(
+        rels_p.as_str(),
+        rels_p.extract_span(),
+    )))
+}
+
+/// Parse `::rename <old> -> <new> [, ...]`.
+fn parse_rename_op(inner: crate::parse::Pair<'_>) -> Result<SysOp> {
+    let mut rename_pairs = vec![];
+    for pair in inner.into_inner() {
+        let mut src = pair.into_inner();
+        let rels_p = src.next().ok_or_else(|| {
+            InvalidQuerySnafu {
+                message: "expected old relation name in rename".to_string(),
+            }
+            .build()
+        })?;
+        let rel = Symbol::new(rels_p.as_str(), rels_p.extract_span());
+        let rels_p = src.next().ok_or_else(|| {
+            InvalidQuerySnafu {
+                message: "expected new relation name in rename".to_string(),
+            }
+            .build()
+        })?;
+        let new_rel = Symbol::new(rels_p.as_str(), rels_p.extract_span());
+        rename_pairs.push((rel, new_rel));
+    }
+    Ok(SysOp::RenameRelation(rename_pairs))
+}
+
+/// Parse `::access_level <level> <relation> [, ...]`.
+fn parse_access_level_op(inner: crate::parse::Pair<'_>) -> Result<SysOp> {
+    let mut ps = inner.into_inner();
+    let access_level = match ps
+        .next()
+        .ok_or_else(|| {
+            InvalidQuerySnafu {
+                message: "expected access level".to_string(),
+            }
+            .build()
+        })?
+        .as_str()
+    {
+        "normal" => AccessLevel::Normal,
+        "protected" => AccessLevel::Protected,
+        "read_only" => AccessLevel::ReadOnly,
+        "hidden" => AccessLevel::Hidden,
+        s => {
+            return Err(InvalidQuerySnafu {
+                message: format!("unexpected access level: {}", s),
+            }
+            .build()
+            .into());
+        }
+    };
+    let rels: Vec<_> = ps
+        .map(|rel_p| Symbol::new(rel_p.as_str(), rel_p.extract_span()))
+        .collect();
+    Ok(SysOp::SetAccessLevel(rels, access_level))
+}
+
+/// Parse `::set_triggers <relation> { on put { ... } on rm { ... } ... }`.
+fn parse_trigger_op(
+    inner: crate::parse::Pair<'_>,
+    _param_pool: &BTreeMap<String, DataValue>,
+    algorithms: &BTreeMap<String, Arc<Box<dyn FixedRule>>>,
+    cur_vld: ValidityTs,
+) -> Result<SysOp> {
+    let mut src = inner.into_inner();
+    let rels_p = src.next().ok_or_else(|| {
+        InvalidQuerySnafu {
+            message: "expected relation name".to_string(),
+        }
+        .build()
+    })?;
+    let rel = Symbol::new(rels_p.as_str(), rels_p.extract_span());
+    let mut puts = vec![];
+    let mut rms = vec![];
+    let mut replaces = vec![];
+    for clause in src {
+        let mut clause_inner = clause.into_inner();
+        let op = clause_inner.next().ok_or_else(|| {
+            InvalidQuerySnafu {
+                message: "expected trigger operation".to_string(),
+            }
+            .build()
+        })?;
+        let script = clause_inner.next().ok_or_else(|| {
+            InvalidQuerySnafu {
+                message: "expected trigger script".to_string(),
+            }
+            .build()
+        })?;
+        let script_str = script.as_str();
+        // Validate the trigger script by parsing it (result is discarded).
+        parse_query(
+            script.into_inner(),
+            &Default::default(),
+            algorithms,
+            cur_vld,
+        )?;
+        match op.as_rule() {
+            Rule::trigger_put => puts.push(script_str.to_string()),
+            Rule::trigger_rm => rms.push(script_str.to_string()),
+            Rule::trigger_replace => replaces.push(script_str.to_string()),
+            r => {
+                return Err(InvalidQuerySnafu {
+                    message: format!("unexpected rule {:?} in trigger parser", r),
+                }
+                .build()
+                .into());
+            }
+        }
+    }
+    Ok(SysOp::SetTriggers(rel, puts, rms, replaces))
+}
+
+/// Parse an advanced index operation (FTS, HNSW, or LSH) that uses
+/// `index_create_adv` or `index_drop` syntax.
+///
+/// The `create_fn` parameter selects which index type to parse.
+fn parse_adv_index_op(
+    inner: crate::parse::Pair<'_>,
+    param_pool: &BTreeMap<String, DataValue>,
+    create_fn: fn(crate::parse::Pair<'_>, &BTreeMap<String, DataValue>) -> Result<SysOp>,
+) -> Result<SysOp> {
+    let inner = inner.into_inner().next().ok_or_else(|| {
+        InvalidQuerySnafu {
+            message: "expected index operation".to_string(),
+        }
+        .build()
+    })?;
+    match inner.as_rule() {
+        Rule::index_create_adv => create_fn(inner, param_pool),
+        Rule::index_drop => parse_index_drop(inner),
+        r => Err(InvalidQuerySnafu {
+            message: format!("unexpected rule {:?} in index parser", r),
+        }
+        .build()
+        .into()),
+    }
+}
+
+/// Parse a standard (B-tree) index operation: `::index create` or `::index drop`.
+fn parse_standard_index_op(inner: crate::parse::Pair<'_>) -> Result<SysOp> {
+    let inner = inner.into_inner().next().ok_or_else(|| {
+        InvalidQuerySnafu {
+            message: "expected index operation".to_string(),
+        }
+        .build()
+    })?;
+    match inner.as_rule() {
+        Rule::index_create => {
+            let _span = inner.extract_span();
+            let mut inner = inner.into_inner();
+            let rel = inner.next().ok_or_else(|| {
+                InvalidQuerySnafu {
+                    message: "expected relation name".to_string(),
+                }
+                .build()
+            })?;
+            let name = inner.next().ok_or_else(|| {
+                InvalidQuerySnafu {
+                    message: "expected index name".to_string(),
+                }
+                .build()
+            })?;
+            let cols: Vec<_> = inner
+                .map(|p| Symbol::new(p.as_str(), p.extract_span()))
+                .collect();
+            if cols.is_empty() {
+                return Err(InvalidQuerySnafu {
+                    message: "index must have at least one column specified".to_string(),
+                }
+                .build()
+                .into());
+            }
+            Ok(SysOp::CreateIndex(
+                Symbol::new(rel.as_str(), rel.extract_span()),
+                Symbol::new(name.as_str(), name.extract_span()),
+                cols,
+            ))
+        }
+        Rule::index_drop => parse_index_drop(inner),
+        r => Err(InvalidQuerySnafu {
+            message: format!("unexpected rule {:?} in index parser", r),
+        }
+        .build()
+        .into()),
+    }
 }
